Fixed compiler warnings.
git-svn-id: svn+ssh://svn.code.sf.net/p/supertuxkart/code/main/trunk@7442 178a84e3-b1eb-0310-8ba1-8eac791a3b58

+++ src/guiengine/skin.cpp
@@ -1273,7 +1273,7 @@ void Skin::renderSections(ptr_vector<Widget>* within_vector)
                 
                 // there's about 40 empty pixels at the top of bar.png
                 ITexture* tex = irr_driver->getTexture( file_manager->getGUIDir() + "bar.png" );
-                irr_driver->getVideoDriver()->draw2DImage(tex, core::rect<s32>(0, widget.m_y - 40*y_size,
+                irr_driver->getVideoDriver()->draw2DImage(tex, core::rect<s32>(0, (int)(widget.m_y - 40*y_size),
                                                                                framesize.Width, framesize.Height),
                                                           core::rect<s32>(core::dimension2di(0,0), tex->getSize()),
                                                           0, 0, true /* alpha */
